% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatParksPalettes.R
\name{colorblind.friendly}
\alias{colorblind.friendly}
\title{Colorblind-Friendly Palette Check}
\usage{
colorblind.friendly(palette_name)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{Acadia}, \code{Arches}, \code{Arches2}, \code{Banff}, \code{Bryce Canyon},
\code{CapitolReef}, \code{Charmonix}, \code{Cuyahoga}, \code{DeathValley}, \code{Denali}, 
\code{Everglades}, \code{Glacier}, \code{GrandCanyon}, \code{Halekala}, \code{IguazuFalls}, 
\code{KingsCanyon}, \code{LakeNakuru}, \code{Olympic}, \code{Redwood}, \code{RockyMtn}, 
\code{Saguaro}, \code{SmokyMtns}, \code{SouthDowns}, \code{Torres}, \code{Triglav}, 
\code{WindCave}, \code{Volcanoes},  \code{Yellowstone}, and \code{Yosemite}.}
}
\value{
TRUE/FALSE value whether palette is colorblind-friendly
}
\description{
Checks whether a palette is colorblind-friendly. Colorblind-friendliness tested using the 'colorblindcheck' package.
To be colorblind-friendly, all colors in the palettes must be distinguishable with deuteranopia, protanopia, and tritanopia.
}
\examples{
colorblind.friendly("DeathValley")
}
