% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatParksPalettes.R
\docType{data}
\name{NatParksPalettes}
\alias{NatParksPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 15.
}
\usage{
NatParksPalettes
}
\description{
Use names(NatParksPalettes) to return all possible palette names. Current choices are:
\code{Acadia}, \code{Arches}, \code{Arches2}, \code{Banff}, \code{Bryce Canyon},
\code{DeathValley}, \code{Everglades}, \code{Glacier}, \code{GrandCanyon}, \code{KingsCanyon},
\code{Olympic}, \code{Redwood}, \code{SmokyMtns}, \code{Yellowstone}, and \code{Yosemite}.
Use \code{\link{natparks.pals}} to construct palettes.
}
\keyword{datasets}
