% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatParksPalettes.R
\name{scale_fill_natparks_c}
\alias{scale_fill_natparks_c}
\title{NatParksPalettes for plotting with ggplot2}
\usage{
scale_fill_natparks_c(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of Palette. Choices are:
\code{Acadia}, \code{Arches}, \code{Arches2}, \code{Banff}, \code{Bryce Canyon},
\code{CapitolReef}, \code{Charmonix}, \code{Cuyahoga}, \code{DeathValley}, \code{Denali}, 
\code{Everglades}, \code{Glacier}, \code{GrandCanyon}, \code{Halekala}, \code{IguazuFalls}, 
\code{KingsCanyon}, \code{LakeNakuru}, \code{Olympic}, \code{Redwood}, \code{RockyMtn}, 
\code{Saguaro}, \code{SmokyMtns}, \code{SouthDowns}, \code{Torres}, \code{Triglav}, 
\code{WindCave}, \code{Volcanoes},  \code{Yellowstone}, and \code{Yosemite}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\value{
A function that returns a continuous color scale.
}
\description{
Function for using \code{NatParksPalettes} colors schemes in \code{ggplot2}. Use \code{\link{scale_color_natparks_d}} and \code{\link{scale_fill_natparks_d}}
for discrete scales and \code{\link{scale_color_natparks_c}} and \code{\link{scale_fill_natparks_c}} for continuous scales.
}
