% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatParksPalettes.R
\docType{data}
\name{NatParksPalettes}
\alias{NatParksPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 30.
}
\usage{
NatParksPalettes
}
\description{
Use names(NatParksPalettes) to return all possible palette names. Current choices are:
\code{Acadia}, \code{Arches}, \code{Arches2}, \code{Banff}, \code{Bryce Canyon},
\code{CapitolReef}, \code{Charmonix}, \code{Cuyahoga}, \code{DeathValley}, \code{Denali}, 
\code{Everglades}, \code{Glacier}, \code{GrandCanyon}, \code{Halekala}, \code{IguazuFalls}, 
\code{KingsCanyon}, \code{LakeNakuru}, \code{Olympic}, \code{Redwood}, \code{RockyMtn}, 
\code{Saguaro}, \code{SmokyMtns}, \code{SouthDowns}, \code{Torres}, \code{Triglav}, 
\code{WindCave}, \code{Volcanoes},  \code{Yellowstone}, and \code{Yosemite}.
Use \code{\link{natparks.pals}} to construct palettes.
}
\keyword{datasets}
